% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_metadata.R
\name{cell_metadata}
\alias{cell_metadata}
\title{Retrieve metadata associated with a cell (cl) used in LINCS experiments.}
\usage{
cell_metadata(cl)
}
\arguments{
\item{cl}{String representng a cell name of interest}
}
\value{
data frame with available cell metadata
}
\description{
Retrieve metadata associated with a cell (cl) used in LINCS experiments.
}
\examples{
cell_metadata("mcf7")
cell_metadata("PC-3")

}
