% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalevels.R
\name{datalevels}
\alias{datalevels}
\title{Retrieve all available data level and associated dataset id for a given dataset id.}
\usage{
datalevels(lds_id)
}
\arguments{
\item{lds_id}{String representng a dataset id (LDS-xxxx)}
}
\value{
data frame with available data level and associated dataset id
}
\description{
Retrieve all available data level and associated dataset id for a given dataset id.
}
\examples{
datalevels("LDS-1233")

}
