% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_entities.R
\name{dataset_entities}
\alias{dataset_entities}
\title{Retrieve entities (gene, protein, small molecule, etc.) associated with a specific LINCS dataset}
\usage{
dataset_entities(lds_id)
}
\arguments{
\item{lds_id}{String representing the dataset id (LDS-xxxx)}
}
\value{
list of vectors with entities, measured or used to perturb the experiment associated to a given dataset. One vector per entity type.
}
\description{
Retrieve entities (gene, protein, small molecule, etc.) associated with a specific LINCS dataset
}
\examples{
dataset_entities("LDS-1266")
dataset_entities("LDS-1173")

}

