% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download LINCS dataset packages, given dataset id's, from Lincs Data Portal.}
\usage{
download_dataset(lds_id, path = "./")
}
\arguments{
\item{lds_id}{Vector of dataset ids (LDS-xxxx)}

\item{path}{Path to local folder where the dataset package will be downloaded, default = current directory}
}
\description{
Download LINCS dataset packages, given dataset id's, from Lincs Data Portal.
}
\examples{
\dontrun{
download_dataset("LDS-1266")
download_dataset("LDS-1173", path = "/datasets")
}

}
