% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dataset.R
\name{find_dataset}
\alias{find_dataset}
\title{Retrieve datasets from LINCS data portal associated with a term of interest (gene, protein, small molecule)}
\usage{
find_dataset(searchterm)
}
\arguments{
\item{searchterm}{String representing the search term}
}
\value{
data.frame containing dataset id and description
}
\description{
Retrieve datasets from LINCS data portal associated with a term of interest (gene, protein, small molecule)
}
\examples{
find_dataset("afatinib")
find_dataset("mcf7")

}
