% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_metadata.R
\name{gene_metadata}
\alias{gene_metadata}
\title{Retrieve metadata associated with a gene measured in LINCS experiments.}
\usage{
gene_metadata(ge)
}
\arguments{
\item{ge}{String representing a gene symbol of interest}
}
\value{
data frame with available gene metadata
}
\description{
Retrieve metadata associated with a gene measured in LINCS experiments.
}
\examples{
gene_metadata("BRCA1")
gene_metadata("CDK1")

}
