% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_metadata.R
\name{protein_metadata}
\alias{protein_metadata}
\title{Retrieve metadata associated with a protein measured or used in LINCS experiments.}
\usage{
protein_metadata(pr)
}
\arguments{
\item{pr}{String representing a protein of interest}
}
\value{
data frame with available protein metadata
}
\description{
Retrieve metadata associated with a protein measured or used in LINCS experiments.
}
\examples{
protein_metadata("ACK")

}

