% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SM_metadata.R
\name{SM_metadata}
\alias{SM_metadata}
\title{Retrieve metadata associated with a Small Molecule (SM) used in LINCS experiments.}
\usage{
SM_metadata(sm)
}
\arguments{
\item{sm}{String representng a small molecule of interest}
}
\value{
data frame with available small molecule metadata
}
\description{
Retrieve metadata associated with a Small Molecule (SM) used in LINCS experiments.
}
\examples{
SM_metadata("afatinib")

}

